Added new entity called Speaker
Here, we've created a new entity (without attributes yet) called Speaker. Take note of the classes that cuba platform generated when we create an entity of type `Persistent. New artifacts are created under the global module, which means cuba platform generates domain model under the global module. So, aside from the Speaker class that was automatically generated and already filled with platform specific annotation, the platform also added an entry under `messages.properties` and an entry into the `persistency.xml` which I suspect is for data persistency purposes.