Added Session entity
Added the Session entity through the cuba studio wizard. So every time we add an entity, cuba platform creates a new class for the entity under the global module which is jpa annotated. It then registers that in persistence.xml and creates entry in messages.properties for the string representation fo the attributes and the class itself.