Initial database user configuration corrections
So some lessons learned here when initialising database users for cuba platform. We need to make sure the database user we will be using has privilege to create new databases (at least). Not so sure if we need to create the database ahead though since cuba studio has a functionality to create the database for us.